New method for defining configuration options and metadata
The additional metadata is used to validate types and ranges of the
configuration values. This is a preliminary step to providing a GUI frontend
to the flame solver.

This also allows the saved configuration (e.g. run/caseName/conf) to contain
only the non-default values specified for a particular simulation.