Strain rate can now be an arbitrary function of time
Any Python function or callable object can be used to define the strain rate. At
the start of each global timestep, a new Chebyshev interpolant is computed,
which is then used to compute a(t) and a'(t). This approach avoids the
complexity and performance penalty of calling the Python function from the C++
solver class.