Check if Cantera supports derived classes of MultiTransport
The changes to make deriving from MultiTransport in a useful way possible won't
be part of a released version of Cantera until 2.1. This enables us to fall back
to the normal version of MultiTransport if we can't use our derived class, which
allows us to keep the required Cantera version at 2.0.

Introduces a "config.h" file populated by SCons.